backend/requirements.txt
EOL
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.0.0
googlemaps==4.10.0
python-decouple==3.8
EOL